Make syncInApp package-private, rewrite a bunch of tests
Also, extract static methods from IterableNotificationBuilder into IterableNotificationHelper to make them testable
