+++ iterableapi/build.gradle
@@ -33,15 +33,15 @@ dependencies {
     api 'com.google.firebase:firebase-messaging:15.0.2'
 
     testImplementation 'junit:junit:4.12'
-    testImplementation 'org.mockito:mockito-core:2.23.0'
+    testImplementation 'org.mockito:mockito-core:2.27.0'
     testImplementation 'org.robolectric:robolectric:4.2.1'
     testImplementation 'org.robolectric:shadows-supportv4:4.2.1'
     testImplementation 'org.robolectric:shadows-playservices:4.2.1'
     testImplementation 'org.khronos:opengl-api:gl1.1-android-2.1_r1'
-    testImplementation "org.powermock:powermock-module-junit4:2.0.0"
-    testImplementation "org.powermock:powermock-module-junit4-rule:2.0.0"
-    testImplementation "org.powermock:powermock-api-mockito2:2.0.0"
-    testImplementation "org.powermock:powermock-classloading-xstream:2.0.0"
+    testImplementation "org.powermock:powermock-module-junit4:2.0.2"
+    testImplementation "org.powermock:powermock-module-junit4-rule:2.0.2"
+    testImplementation "org.powermock:powermock-api-mockito2:2.0.2"
+    testImplementation "org.powermock:powermock-classloading-xstream:2.0.2"
     testImplementation 'com.squareup.okhttp3:mockwebserver:3.10.0'
     testImplementation 'org.skyscreamer:jsonassert:1.5.0'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
