+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableInAppManager.java
@@ -85,7 +85,7 @@ public class IterableInAppManager implements IterableActivityMonitor.AppStateCal
      * Trigger a manual sync. This method is called automatically by the SDK, so there should be no
      * need to call this method from your app.
      */
-    public void syncInApp() {
+    void syncInApp() {
         this.api.getInAppMessages(MESSAGES_TO_FETCH, new IterableHelper.IterableActionHandler() {
             @Override
             public void execute(String payload) {
